+++ app/loading.tsx
@@ -1,0 +1,9 @@
+import Container from '../components/Container'
+
+export default function Loading() {
+  return (
+    <Container>
+      <div className='py-10 text-center'>Loading ...</div>
+    </Container>
+  )
+}
