+++ components/ThemeProvider.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { ThemeProvider } from 'next-themes'
+
+export default function Providers({ children }: { children: React.ReactNode }) {
+  return <ThemeProvider attribute='class'>{children}</ThemeProvider>
+}
