+++ interfaces/comment.ts
@@ -14,14 +14,14 @@ export type Comment = {
 }
 
 export type Post = {
-  slug?: string
-  title?: string
+  slug: string
+  title: string
   author?: string
-  date?: Date
-  content?: string
+  date: Date
+  content: string
   excerpt?: string
-  category?: string
-  category_slug?: string
+  category: string
+  category_slug: string
   tags?: string[]
   edit_url?: string
   [key: string]: any
