+++ lib/comment/deleteComment.ts
@@ -8,7 +8,7 @@ export default async function deleteComments(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const url = clearUrl(req.headers.referer)
+  const url = clearUrl(req.headers.referer || '')
   const { comment }: { url: string; comment: Comment } = req.body
   const { authorization } = req.headers
