+++ package.json
@@ -20,7 +20,7 @@
     "gray-matter": "4.0.3",
     "lucide-react": "^0.209.0",
     "nanoid": "^4.0.0",
-    "next": "latest",
+    "next": "^13.4.1",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -44,7 +44,7 @@
     "@types/rss": "^0.0.30",
     "autoprefixer": "^10.4.12",
     "eslint": "8.39.0",
-    "eslint-config-next": "13.4.0",
+    "eslint-config-next": "^13.4.1",
     "postcss": "^8.4.18",
     "prettier": "^2.8.7",
     "tailwindcss": "^3.2.1",
